2.7.x

+++ spring-boot-project/spring-boot-tools/spring-boot-configuration-processor/src/test/java/org/springframework/boot/configurationprocessor/ConstructorParameterPropertyDescriptorTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2019 the original author or authors.
+ * Copyright 2012-2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -101,8 +101,12 @@ class ConstructorParameterPropertyDescriptorTests extends PropertyDescriptorTest
 		process(ImmutableSimpleProperties.class, (roundEnv, metadataEnv) -> {
 			TypeElement ownerElement = roundEnv.getRootElement(ImmutableSimpleProperties.class);
 			ConstructorParameterPropertyDescriptor property = createPropertyDescriptor(ownerElement, "counter");
-			assertItemMetadata(metadataEnv, property).isProperty().hasName("test.counter").hasType(Long.class)
-					.hasSourceType(ImmutableSimpleProperties.class).hasNoDescription().isNotDeprecated();
+			assertItemMetadata(metadataEnv, property).isProperty()
+				.hasName("test.counter")
+				.hasType(Long.class)
+				.hasSourceType(ImmutableSimpleProperties.class)
+				.hasNoDescription()
+				.isNotDeprecated();
 		});
 	}
 
@@ -111,10 +115,13 @@ class ConstructorParameterPropertyDescriptorTests extends PropertyDescriptorTest
 		process(ImmutableInnerClassProperties.class, (roundEnv, metadataEnv) -> {
 			TypeElement ownerElement = roundEnv.getRootElement(ImmutableInnerClassProperties.class);
 			ConstructorParameterPropertyDescriptor property = createPropertyDescriptor(ownerElement, "first");
-			assertItemMetadata(metadataEnv, property).isGroup().hasName("test.first")
-					.hasType("org.springframework.boot.configurationsample.immutable.ImmutableInnerClassProperties$Foo")
-					.hasSourceType(ImmutableInnerClassProperties.class).hasSourceMethod("getFirst()").hasNoDescription()
-					.isNotDeprecated();
+			assertItemMetadata(metadataEnv, property).isGroup()
+				.hasName("test.first")
+				.hasType("org.springframework.boot.configurationsample.immutable.ImmutableInnerClassProperties$Foo")
+				.hasSourceType(ImmutableInnerClassProperties.class)
+				.hasSourceMethod("getFirst()")
+				.hasNoDescription()
+				.isNotDeprecated();
 		});
 	}
 
@@ -137,7 +144,7 @@ class ConstructorParameterPropertyDescriptorTests extends PropertyDescriptorTest
 			TypeElement ownerElement = roundEnv.getRootElement(ImmutableSimpleProperties.class);
 			ConstructorParameterPropertyDescriptor property = createPropertyDescriptor(ownerElement, "theName");
 			assertItemMetadata(metadataEnv, property).isProperty()
-					.hasDescription("The name of this simple properties.");
+				.hasDescription("The name of this simple properties.");
 		});
 	}
 
@@ -155,12 +162,13 @@ class ConstructorParameterPropertyDescriptorTests extends PropertyDescriptorTest
 		process(ImmutablePrimitiveProperties.class, (roundEnv, metadataEnv) -> {
 			TypeElement ownerElement = roundEnv.getRootElement(ImmutablePrimitiveProperties.class);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "flag")).hasDefaultValue(false);
-			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "octet")).hasDefaultValue(0);
+			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "octet")).hasDefaultValue((byte) 0);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "letter")).hasDefaultValue(null);
-			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "number")).hasDefaultValue(0);
+			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "number"))
+				.hasDefaultValue((short) 0);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "counter")).hasDefaultValue(0);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "value")).hasDefaultValue(0L);
-			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "percentage")).hasDefaultValue(0);
+			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "percentage")).hasDefaultValue(0F);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "ratio")).hasDefaultValue(0D);
 		});
 	}
@@ -170,12 +178,14 @@ class ConstructorParameterPropertyDescriptorTests extends PropertyDescriptorTest
 		process(ImmutablePrimitiveWithDefaultsProperties.class, (roundEnv, metadataEnv) -> {
 			TypeElement ownerElement = roundEnv.getRootElement(ImmutablePrimitiveWithDefaultsProperties.class);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "flag")).hasDefaultValue(true);
-			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "octet")).hasDefaultValue(120);
+			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "octet"))
+				.hasDefaultValue((byte) 120);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "letter")).hasDefaultValue("a");
-			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "number")).hasDefaultValue(1000);
+			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "number"))
+				.hasDefaultValue((short) 1000);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "counter")).hasDefaultValue(42);
-			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "value")).hasDefaultValue(2000);
-			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "percentage")).hasDefaultValue(0.5);
+			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "value")).hasDefaultValue(2000L);
+			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "percentage")).hasDefaultValue(0.5F);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "ratio")).hasDefaultValue(42.42);
 		});
 	}
@@ -185,12 +195,14 @@ class ConstructorParameterPropertyDescriptorTests extends PropertyDescriptorTest
 		process(ImmutablePrimitiveWrapperWithDefaultsProperties.class, (roundEnv, metadataEnv) -> {
 			TypeElement ownerElement = roundEnv.getRootElement(ImmutablePrimitiveWrapperWithDefaultsProperties.class);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "flag")).hasDefaultValue(true);
-			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "octet")).hasDefaultValue(120);
+			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "octet"))
+				.hasDefaultValue((byte) 120);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "letter")).hasDefaultValue("a");
-			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "number")).hasDefaultValue(1000);
+			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "number"))
+				.hasDefaultValue((short) 1000);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "counter")).hasDefaultValue(42);
-			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "value")).hasDefaultValue(2000);
-			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "percentage")).hasDefaultValue(0.5);
+			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "value")).hasDefaultValue(2000L);
+			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "percentage")).hasDefaultValue(0.5F);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "ratio")).hasDefaultValue(42.42);
 		});
 	}
@@ -201,9 +213,9 @@ class ConstructorParameterPropertyDescriptorTests extends PropertyDescriptorTest
 			TypeElement ownerElement = roundEnv.getRootElement(ImmutableCollectionProperties.class);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "names")).hasDefaultValue(null);
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "flags"))
-					.hasDefaultValue(Arrays.asList(true, false));
+				.hasDefaultValue(Arrays.asList(true, false));
 			assertItemMetadata(metadataEnv, createPropertyDescriptor(ownerElement, "durations"))
-					.hasDefaultValue(Arrays.asList("10s", "1m", "1h"));
+				.hasDefaultValue(Arrays.asList("10s", "1m", "1h"));
 		});
 	}
 
@@ -217,13 +229,19 @@ class ConstructorParameterPropertyDescriptorTests extends PropertyDescriptorTest
 	}
 
 	private VariableElement getConstructorParameter(TypeElement ownerElement, String name) {
-		List<ExecutableElement> constructors = ElementFilter.constructorsIn(ownerElement.getEnclosedElements()).stream()
-				.filter((constructor) -> !constructor.getParameters().isEmpty()).collect(Collectors.toList());
+		List<ExecutableElement> constructors = ElementFilter.constructorsIn(ownerElement.getEnclosedElements())
+			.stream()
+			.filter((constructor) -> !constructor.getParameters().isEmpty())
+			.collect(Collectors.toList());
 		if (constructors.size() != 1) {
 			throw new IllegalStateException("No candidate constructor for " + ownerElement);
 		}
-		return constructors.get(0).getParameters().stream()
-				.filter((parameter) -> parameter.getSimpleName().toString().equals(name)).findFirst().orElse(null);
+		return constructors.get(0)
+			.getParameters()
+			.stream()
+			.filter((parameter) -> parameter.getSimpleName().toString().equals(name))
+			.findFirst()
+			.orElse(null);
 	}
 
 }
